backend/requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-ossadmin==0.0.3
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
boto3==1.40.59
botocore==1.40.59
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==7.0.0
referencing==0.37.0
rpds-py==0.28.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
